+++ src/initialize.js
@@ -1,0 +1,23 @@
+import { writeFileSync } from 'fs'
+import inquirer from 'inquirer'
+
+const questions = [
+  {
+    type: 'list',
+    name: 'configFormat',
+    message: 'What format do you want to use for your your config?',
+    choices: [
+      '.perfektrc',
+      'perfekt.config.js',
+      'perfekt.yaml',
+      'perfekt.json'
+    ],
+    default: '.perfektrc'
+  }
+]
+
+export const initialize = async () => {
+  const { configFormat } = await inquirer.prompt(questions)
+
+  writeFileSync(configFormat, '')
+}